Update the all pages using the NewsCategory
